refactoring: moved useGetRBTCPrice to general hooks
