+++ apps/frontend/src/hooks/zero/useGetRBTCPrice.ts
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 import { EthersLiquity, ReadableEthersLiquity } from '@sovryn-zero/lib-ethers';
 import { getProvider } from '@sovryn/ethers-provider';
 
-import { getRskChainId } from '../../../../utils/chain';
+import { getRskChainId } from '../../utils/chain';
 
 export const useGetRBTCPrice = () => {
   const [price, setPrice] = useState('0');
